dockerized proteogenomics pipeline

+++ proteogenomics/bin/python/mq_genome_to_peptide.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 import pandas as pd
 import sys
@@ -9,19 +9,23 @@ import shutil
 import os
 from collections import defaultdict
 import pickle
+import yaml
 
-loader = importlib.machinery.SourceFileLoader('config', sys.argv[1])
-config = loader.load_module()
+config = yaml.load(open(sys.argv[1]))
 
 output = sys.argv[2]
 
-peptides = pd.read_csv(config.mq_txt + 'peptides.txt', sep='\t',engine='python')
+os.mkdir(output +'/strains')
 
-peptides = peptides[(peptides['Potential contaminant'] != '+') & (peptides['Reverse'] != '+')]
+peptides = pd.read_csv(config['mq_txt'] + '/peptides.txt', sep='\t',engine='python')
+
+peptides = peptides[(peptides['Potential contaminant'].isnull()) & (peptides['Reverse'].isnull())]
+
+all_peptides = peptides['Sequence'].tolist()
 
 strain_samples = defaultdict(list)
 
-samples = config.samples
+samples = config['samples']
 
 for sample in samples:
     st = samples[sample]['STRAIN']
@@ -35,7 +39,13 @@ exp_cols = [i for i in peptides.columns if i.startswith('Experiment')]
 #assert 4 == 5
 # Specific peptides for each strain
 
-for strain in config.strains:
+def check_identified(peptide, identified):
+    if peptide in identified:
+        return '+'
+    else:
+        return '-'
+
+for strain in config['strains']:
     samples = strain_samples[strain]
     sample_columns = []
 
@@ -44,19 +54,25 @@ for strain in config.strains:
 
     strain_filt = peptides[peptides[sample_columns].sum(axis=1) >=1]
     strain_peptides = strain_filt['Sequence'].tolist()
-    paths = config.strains[strain]
-    genome = list(SeqIO.parse(paths['sf_genome'],'fasta'))
-    g2p = sequtils.peptides2genome(genome, assembly_name = str(strain), translation_table=config.translation_table, peptides_list=strain_peptides, threads=config.threads)
-    speps = g2p.peptides
-    strainpath=output +'/' + strain
-    try:
-        shutil.rmtree(strainpath)
-        os.mkdir(strainpath)
-    except:
-        os.mkdir(strainpath)    
-    outpath=strainpath + '/' + '{}_mapped_peptides.p'.format(str(strain))
-    #speps.to_csv(strainpath + '/' + '{}_mapped_peptides.csv'.format(str(strain)))
-    pickle.dump( speps, open( outpath, "wb" ) )
+    paths = config['strains'][strain]
+    
+    if paths['sf_genome'] != None:
+        print(strain)
+        genome = list(SeqIO.parse(paths['sf_genome'],'fasta'))
+
+        g2p = sequtils.peptides2genome(genome, assembly_name = str(strain), translation_table=config['translation_table'], peptides_list=all_peptides, threads=config['threads'])
+        speps = g2p.peptides
+        speps['Strain_identified'] = speps['Peptide_sequence'].apply(lambda x : check_identified(x, strain_peptides))
+        strainpath=output +'/strains/' + strain
+        try:
+            shutil.rmtree(strainpath)
+            os.mkdir(strainpath)
+        except:
+            os.mkdir(strainpath)    
+        
+        outpath=strainpath + '/' + '{}_mapped_peptides.p'.format(str(strain))   
+        #speps.to_csv(strainpath + '/' + '{}_mapped_peptides.csv'.format(str(strain)))
+        pickle.dump( speps, open( outpath, "wb" ) )
